+++ src/index.ts
@@ -1,5 +1,7 @@
 import Printify from './client';
 
+export type { PrintifyError } from './http';
+
 export * from './v1/types';
 export * from './v2/types';